Check user rights on updating methods (#768)

+++ Car.Backend/Car.Domain/Services/Implementation/JourneyService.cs
@@ -171,13 +171,13 @@ namespace Car.Domain.Services.Implementation
 
             var date = DateTime.Now.AddDays(13);
 
-            foreach (var schedule in schedules)
+            schedules.ForEach(async (schedule) =>
             {
                 if (schedule.Days.ToString().Contains(date.DayOfWeek.ToString()))
                 {
                     await AddFutureJourneyAsync(schedule, date);
                 }
-            }
+            });
         }
 
         public async Task<JourneyModel> AddJourneyAsync(JourneyDto journeyModel) => await AddJourneyAsync(journeyModel, null);
@@ -215,7 +215,7 @@ namespace Car.Domain.Services.Implementation
             return true;
         }
 
-        public async Task CancelAsync(int journeyId)
+        public async Task<bool> CancelAsync(int journeyId)
         {
             var journeyToCancel = await journeyRepository
                 .Query()
@@ -226,6 +226,13 @@ namespace Car.Domain.Services.Implementation
 
             if (journeyToCancel is not null)
             {
+                int userId = httpContextAccessor.HttpContext!.User.GetCurrentUserId();
+
+                if (userId != journeyToCancel.OrganizerId)
+                {
+                    return false;
+                }
+
                 journeyToCancel.IsCancelled = true;
                 await journeyRepository.SaveChangesAsync();
                 await notificationService.DeleteNotificationsAsync(journeyToCancel.Notifications);
@@ -244,13 +251,17 @@ namespace Car.Domain.Services.Implementation
                     }
                 }
             }
+
+            return true;
         }
 
-        public async Task<JourneyModel?> UpdateRouteAsync(JourneyDto journeyDto) => await UpdateRouteAsync(journeyDto, false);
+        public async Task<(bool IsUpdated, JourneyModel? UpdatedJourney)> UpdateRouteAsync(JourneyDto journeyDto)
+            => await UpdateRouteAsync(journeyDto, false);
 
-        public async Task<JourneyModel?> UpdateDetailsAsync(JourneyDto journeyDto) => await UpdateDetailsAsync(journeyDto, false);
+        public async Task<(bool IsUpdated, JourneyModel? UpdatedJourney)> UpdateDetailsAsync(JourneyDto journeyDto)
+            => await UpdateDetailsAsync(journeyDto, false);
 
-        public async Task<InvitationDto> UpdateInvitationAsync(InvitationDto invitationDto)
+        public async Task<(bool IsUpdated, InvitationDto? UpdatedInvitationDto)> UpdateInvitationAsync(InvitationDto invitationDto)
         {
             var journey = await journeyRepository.Query()
                 .IncludeJourneyInvitations()
@@ -259,13 +270,20 @@ namespace Car.Domain.Services.Implementation
 
             if (invitation != null)
             {
+                int userId = httpContextAccessor.HttpContext!.User.GetCurrentUserId();
+
+                if (userId != invitation.InvitedUserId)
+                {
+                    return (false, null);
+                }
+
                 invitation.Type = invitationDto.Type;
 
                 invitation = await invitationRepository.UpdateAsync(invitation);
                 await journeyRepository.SaveChangesAsync();
             }
 
-            return mapper.Map<Invitation, InvitationDto>(invitation!);
+            return (true, mapper.Map<Invitation, InvitationDto>(invitation!));
         }
 
         public IEnumerable<ApplicantJourney> GetApplicantJourneys(JourneyFilter filter)
@@ -345,7 +363,7 @@ namespace Car.Domain.Services.Implementation
             return true;
         }
 
-        public async Task<bool> AddUserToJourney(JourneyApplyModel journeyApply)
+        public async Task<(bool IsAddingAllowed, bool IsUserAdded)> AddUserToJourney(JourneyApplyModel journeyApply)
         {
             var journeyId = journeyApply.JourneyUser!.JourneyId;
 
@@ -366,7 +384,14 @@ namespace Car.Domain.Services.Implementation
                 || journey.Participants.Contains(userToAdd)
                 || !IsSuitableSeatsCountForAdding(journey, journeyApply.JourneyUser))
             {
-                return false;
+                return (true, false);
+            }
+
+            int currentUserId = httpContextAccessor.HttpContext!.User.GetCurrentUserId();
+
+            if (currentUserId != journey.OrganizerId && currentUserId != journeyApply.JourneyUser.UserId)
+            {
+                return (false, false);
             }
 
             if (journey.Chat is not null)
@@ -399,7 +424,7 @@ namespace Car.Domain.Services.Implementation
                 await journeyUserService.UpdateJourneyUserAsync(journeyApply.JourneyUser);
             }
 
-            return true;
+            return (true, true);
         }
 
         public async Task<int> GetUnreadMessagesCountForNewUserAsync(int journeyId)
@@ -585,7 +610,7 @@ namespace Car.Domain.Services.Implementation
             return mapper.Map<Journey, JourneyModel>(addedJourney!);
         }
 
-        private async Task<JourneyModel?> UpdateRouteAsync(JourneyDto journeyDto, bool isParentUpdated)
+        private async Task<(bool IsUpdated, JourneyModel? UpdatedJourney)> UpdateRouteAsync(JourneyDto journeyDto, bool isParentUpdated)
         {
             var journey = await journeyRepository.Query()
                 .FilterEditable()
@@ -593,45 +618,50 @@ namespace Car.Domain.Services.Implementation
                 .IncludeJourneyPoints()
                 .FirstOrDefaultAsync(j => j.Id == journeyDto.Id);
 
-            if (journey is null)
+            if (journey != null)
             {
-                return null;
+                int userId = httpContextAccessor.HttpContext!.User.GetCurrentUserId();
+
+                if (userId != journey.OrganizerId)
+                {
+                    return (false, null);
+                }
+
+                if (!isParentUpdated)
+                {
+                    journey.ParentId = null;
+                }
+
+                var updatedJourney = mapper.Map<JourneyDto, Journey>(journeyDto);
+
+                journey.Duration = updatedJourney.Duration;
+                journey.Stops = updatedJourney.Stops;
+                journey.JourneyPoints = updatedJourney.JourneyPoints;
+                journey.RouteDistance = updatedJourney.RouteDistance;
+
+                await journeyRepository.SaveChangesAsync();
+
+                if (await scheduleRepository.Query().AnyAsync(schedule => schedule.Id == journeyDto.Id))
+                {
+                    await UpdateChildRoutesAsync(journeyDto);
+                }
+                else if (journeyDto.WeekDay is not null)
+                {
+                    var schedule = await scheduleRepository.AddAsync(new Schedule
+                    { Id = journeyDto.Id, Days = (WeekDays)journeyDto.WeekDay });
+                    await scheduleRepository.SaveChangesAsync();
+                    await AddFutureJourneysAsync(schedule);
+                }
+                else
+                {
+                    await notificationService.JourneyUpdateNotifyUserAsync(await journeyRepository
+                        .Query()
+                        .IncludeAllParticipants()
+                        .FirstOrDefaultAsync(j => j.Id == journeyDto.Id));
+                }
             }
 
-            if (!isParentUpdated)
-            {
-                journey.ParentId = null;
-            }
-
-            var updatedJourney = mapper.Map<JourneyDto, Journey>(journeyDto);
-
-            journey.Duration = updatedJourney.Duration;
-            journey.Stops = updatedJourney.Stops;
-            journey.JourneyPoints = updatedJourney.JourneyPoints;
-            journey.RouteDistance = updatedJourney.RouteDistance;
-
-            await journeyRepository.SaveChangesAsync();
-
-            if (await scheduleRepository.Query().AnyAsync(schedule => schedule.Id == journeyDto.Id))
-            {
-                await UpdateChildRoutesAsync(journeyDto);
-            }
-            else if (journeyDto.WeekDay is not null)
-            {
-                var schedule = await scheduleRepository.AddAsync(new Schedule
-                { Id = journeyDto.Id, Days = (WeekDays)journeyDto.WeekDay });
-                await scheduleRepository.SaveChangesAsync();
-                await AddFutureJourneysAsync(schedule);
-            }
-            else
-            {
-                await notificationService.JourneyUpdateNotifyUserAsync(await journeyRepository
-                    .Query()
-                    .IncludeAllParticipants()
-                    .FirstOrDefaultAsync(j => j.Id == journeyDto.Id));
-            }
-
-            return mapper.Map<Journey, JourneyModel>(journey);
+            return (true, mapper.Map<Journey, JourneyModel>(journey!));
         }
 
         private async Task UpdateChildRoutesAsync(JourneyDto journeyDto)
@@ -666,7 +696,7 @@ namespace Car.Domain.Services.Implementation
             }
         }
 
-        private async Task<JourneyModel?> UpdateDetailsAsync(JourneyDto journeyDto, bool isParentUpdated)
+        private async Task<(bool IsUpdated, JourneyModel? UpdatedJourney)> UpdateDetailsAsync(JourneyDto journeyDto, bool isParentUpdated)
         {
             var journey = await journeyRepository.Query()
                 .IncludeJourneyInvitations()
@@ -674,46 +704,50 @@ namespace Car.Domain.Services.Implementation
                 .FilterEditable()
                 .FirstOrDefaultAsync(j => j.Id == journeyDto.Id);
 
-            if (journey is null)
-            {
-                return null;
-            }
-
-            var existingInvitations = journey.Invitations;
-
             var updatedJourney = mapper.Map<JourneyDto, Journey>(journeyDto);
 
-            if (isParentUpdated)
+            if (journey != null)
             {
-                updatedJourney.ParentId = journey.ParentId;
+                var existingInvitations = journey.Invitations;
+                int userId = httpContextAccessor.HttpContext!.User.GetCurrentUserId();
+
+                if (userId != updatedJourney.OrganizerId)
+                {
+                    return (false, null);
+                }
+
+                if (isParentUpdated)
+                {
+                    updatedJourney.ParentId = journey.ParentId;
+                }
+
+                updatedJourney = await journeyRepository.UpdateAsync(updatedJourney);
+                await journeyRepository.SaveChangesAsync();
+
+                if (await scheduleRepository.Query().AnyAsync(schedule => schedule.Id == journeyDto.Id))
+                {
+                    await UpdateChildDetailsAsync(journeyDto);
+                }
+                else if (journeyDto.WeekDay is not null)
+                {
+                    var schedule = await scheduleRepository.AddAsync(new Schedule
+                    { Id = journeyDto.Id, Days = (WeekDays)journeyDto.WeekDay });
+                    await scheduleRepository.SaveChangesAsync();
+                    await AddFutureJourneysAsync(schedule);
+                }
+                else
+                {
+                    await notificationService.JourneyUpdateNotifyUserAsync(await journeyRepository
+                        .Query()
+                        .IncludeAllParticipants()
+                        .FirstOrDefaultAsync(j => j.Id == updatedJourney.Id));
+
+                    var newInvitations = updatedJourney.Invitations.Except(existingInvitations).ToList();
+                    await NotifyInvitedUsers(newInvitations, updatedJourney.OrganizerId, updatedJourney.Id);
+                }
             }
 
-            updatedJourney = await journeyRepository.UpdateAsync(updatedJourney);
-            await journeyRepository.SaveChangesAsync();
-
-            if (await scheduleRepository.Query().AnyAsync(schedule => schedule.Id == journeyDto.Id))
-            {
-                await UpdateChildDetailsAsync(journeyDto);
-            }
-            else if (journeyDto.WeekDay is not null)
-            {
-                var schedule = await scheduleRepository.AddAsync(new Schedule
-                { Id = journeyDto.Id, Days = (WeekDays)journeyDto.WeekDay });
-                await scheduleRepository.SaveChangesAsync();
-                await AddFutureJourneysAsync(schedule);
-            }
-            else
-            {
-                await notificationService.JourneyUpdateNotifyUserAsync(await journeyRepository
-                    .Query()
-                    .IncludeAllParticipants()
-                    .FirstOrDefaultAsync(j => j.Id == updatedJourney.Id));
-
-                var newInvitations = updatedJourney.Invitations.Except(existingInvitations).ToList();
-                await NotifyInvitedUsers(newInvitations, updatedJourney.OrganizerId, updatedJourney.Id);
-            }
-
-            return mapper.Map<Journey, JourneyModel>(updatedJourney!);
+            return (true, mapper.Map<Journey, JourneyModel>(updatedJourney!));
         }
 
         private async Task UpdateChildDetailsAsync(JourneyDto journeyDto)
